+++ Autopilot/GUI/TerminUebersicht.xaml
@@ -1,0 +1,45 @@
+﻿<Page x:Class="Autopilot.GUI.TerminUebersicht"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      mc:Ignorable="d" 
+      d:DesignHeight="500" d:DesignWidth="600"
+	Title="TerminUebersicht" Loaded="Page_Loaded">    
+    <Grid Margin="0,0,0,0">
+        <DatePicker x:Name="dp_EndeFilter" HorizontalAlignment="Right" Margin="461,10,10,0" VerticalAlignment="Top" SelectedDateChanged="dp_BeginnEndeFilter_SelectedDateChanged"/>
+        <DatePicker x:Name="dp_BeginnFilter" HorizontalAlignment="Right" Margin="0,10,174,0" VerticalAlignment="Top" SelectedDateChanged="dp_BeginnEndeFilter_SelectedDateChanged"/>
+        <Label x:Name="lb_bis" Content="bis" HorizontalAlignment="Right" Margin="0,8,144,0" VerticalAlignment="Top" RenderTransformOrigin="2.329,-1.051"/>
+        <Label x:Name="lb_von" Content="von" HorizontalAlignment="Right" Margin="0,9,308,0" VerticalAlignment="Top"/>
+        <TextBox x:Name="tb_Filter" Height="23" VerticalAlignment="Top" Margin="10,10,340,0" TextWrapping="Wrap" TextChanged="tb_Filter_TextChanged" GotFocus="tb_Filter_GotFocus"/>
+        <DataGrid x:Name="DataGridFilter" AutoGenerateColumns="False" IsReadOnly="True" Margin="10,41,10,240" SelectionChanged="DataGridFilter_SelectionChanged">
+            <DataGrid.Columns>
+                <DataGridTextColumn Binding="{Binding ter_id}" Width="Auto" Header="ter_id" Visibility="Hidden"/>
+                <DataGridTextColumn Binding="{Binding ID_spez}" Width="Auto" Header="IDspez" Visibility="Hidden"/>
+                <DataGridTextColumn Binding="{Binding tart_id}" Width="Auto" Header="tart_id" Visibility="Hidden"/>
+                <DataGridTextColumn Binding="{Binding f_p}" Width="Auto" Header="f_p" Visibility="Hidden"/>
+                <DataGridTextColumn Binding="{Binding tart_bez}" Width="Auto" Header="Terminart"/>
+                <DataGridTextColumn Binding="{Binding ter_beginn}" Width="Auto" Header="von"/>
+                <DataGridTextColumn Binding="{Binding ter_ende}" Width="Auto" Header="bis"/>
+                <DataGridTextColumn Binding="{Binding ter_txt}" Width="Auto" Header="Termin"/>
+            </DataGrid.Columns>
+        </DataGrid>
+        <Label Content="Terminart" HorizontalAlignment="Left" Margin="19,0,0,148" VerticalAlignment="Bottom"/>
+        <ComboBox x:Name="cb_Terminart" HorizontalAlignment="Left" Margin="85,0,0,151" VerticalAlignment="Bottom" Width="160" ItemsSource="{Binding}" SelectedValuePath="tart_id" DisplayMemberPath="tart_bez"/>
+        <Label Content="von" HorizontalAlignment="Left" Margin="51,0,0,121" VerticalAlignment="Bottom" RenderTransformOrigin="0.159,0.487"/>
+        <Label Content="bis" HorizontalAlignment="Left" Margin="55,0,0,91" VerticalAlignment="Bottom" RenderTransformOrigin="-0.158,0.513"/>
+        <DatePicker x:Name="dp_BeginnAuswahl" HorizontalAlignment="Left" Margin="85,0,0,121" VerticalAlignment="Bottom" SelectedDateChanged="dp_BeginnEndeAuswahl_SelectedDateChanged" Width="160"/>
+        <DatePicker x:Name="dp_EndeAuswahl" HorizontalAlignment="Left" Margin="85,0,0,92" VerticalAlignment="Bottom" SelectedDateChanged="dp_BeginnEndeAuswahl_SelectedDateChanged" Width="160"/>
+        <DataGrid x:Name="DataGridInfo" AutoGenerateColumns="False" IsReadOnly="True" Margin="287,10,10,10" VerticalAlignment="Bottom" Height="224">
+            <DataGrid.Columns>
+                <DataGridTextColumn Binding="{Binding tart_bez}" Width="Auto" Header="Terminart"/>
+                <DataGridTextColumn Binding="{Binding ter_beginn}" Width="Auto" Header="von"/>
+                <DataGridTextColumn Binding="{Binding ter_ende}" Width="Auto" Header="bis"/>
+            </DataGrid.Columns>
+        </DataGrid>
+        <Button x:Name="bt_Speichern" Content="Speichern" HorizontalAlignment="Left" Margin="170,0,0,67" VerticalAlignment="Bottom" Width="75" Click="bt_Speichern_Click"/>
+        <Label x:Name="lb_Termin" Content="" HorizontalAlignment="Left" Margin="85,0,0,179" VerticalAlignment="Bottom" Width="160"/>
+        <Button x:Name="bt_Abbrechen" Content="Abbrechen" HorizontalAlignment="Left" Margin="90,0,0,67" VerticalAlignment="Bottom" Width="75" Click="bt_Abbrechen_Click"/>
+        <Button x:Name="bt_Loeschen" Content="Löschen" HorizontalAlignment="Left" Margin="170,0,0,42" VerticalAlignment="Bottom" Width="75" Click="bt_Loeschen_Click"/>
+    </Grid>
+</Page>
